- Using async/await now - reduced garbage generation
Save time boittleneck? DeltaTime (example sheeps) and worst offender? XMLSpawner
Maybe last commited Savetime should be a could be used to go around deltatime?

As in writing the DateTime when the next thing should happen, then look at the differenz in the Datetime of now and the last save time.
That way it could be restored to the same value without having to rely on delta time that changes each save